Add support for passing options to an install
Install can now be done with:

```
mix igniter.install nerves#runtime=false`
```

Producing:

```
{:nerves, "~> x.y.z", [runtime: false]}
```

To avoid a bunch of fun problems with spaces, commas and such the
format is a query string. Spaces and other nasty characters will need
to be encoded. The encoding exists in the standard library. It will
allow creating Igniter install tasks for Nerves systems and similar
that have some fairly particular options.