fix: OTP 28 Compatibility via removing inflex (#288)
* Remove :inflex, :mix_test_watch from dependency list

* Fix configuration handling in IgniterConfig module

Create configs() function to dynamically inject values that can't be
stored as module constants in OTP 28, and update references to use
this function instead of accessing @configs directly.

* Add Inflex.Pluralize module with OTP/28 compatibility fixes

* Add fixes for drive/drives and movie/movies

* Remove inflex, mix_test_watch from mix.lock

----

Note from Zach:

To the best of my knowledge, my packages with igniters are the only ones using inflex, so I will be publishing fixes for mine as soon as this is released to use this new vendored pluralization logic. Apologies if anyone was relying on this :)